requirements.txt
# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Authentication
python-jose[cryptography]==3.3.0

# Validation & Serialization
pydantic==2.6.1
pydantic-settings==2.1.0

# Utilities
httpx==0.26.0
PyYAML>=6.0.1

# Feishu / Lark
lark-oapi==1.5.3
# feishu_fastapi_sdk：由仓库根目录源包 ``feishu_fastapi_sdk/`` 提供（与 feishu_agent 并列），不再从 Git 安装。

# LLM
anthropic[bedrock]>=0.39

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
